update
あとは当たり判定をとってプレイヤーと矢印が当たったらGameOverにいくようにしたら完成

+++ src/Objects/Player/Player.cpp
@@ -96,10 +96,10 @@ void Player::DrawPlayer()
 		SCREEN_SIZE_Y - PLAYER_MOVE_LIMIT_Y + PLAYER_COLLISION_R,
 		GetColor(0, 0, 5));
 
-	DrawExtendGraph(m_pos.x - PLAYER_COLLISION_R,
-		m_pos.y - PLAYER_COLLISION_R,
-		m_pos.x + PLAYER_COLLISION_R,
-		m_pos.y + PLAYER_COLLISION_R,
+	DrawExtendGraph((int)(m_pos.x - PLAYER_COLLISION_R),
+		(int)(m_pos.y - PLAYER_COLLISION_R),
+		(int)(m_pos.x + PLAYER_COLLISION_R),
+		(int)(m_pos.y + PLAYER_COLLISION_R),
 		m_handle,
 		true);
 }
